Allow configurable runit_service cookbook
In our SmartStack wrapper cookbook, we want to modify the initialization
file so that it loads rbenv first. In order to do this, we need to
modify the sv-synapse-run.erb template beyond what the current
attributes allow.

In order to satisfy this use-case, an attribute allowing an arbitrary
cookbook to provide that template would be helpful.